# sstable: pass KeySchema by pointer